Added PDF.js detection (is built into Firefox version 19 or above)
Currently PDFObject tries to detect PDF.js via a MIME type but FF
doesn't seem to have a way to provide a MIME types variable.

We know for certain PDF.js is built-in to FF >= 19 so adding a check if
the FF version is 19 or greater.

Also added a "class exists" check to prevent multiple entries of the
same class being attached to the pdf container element.